--- transforms/code/comment_cleanser/python/.dockerignore
@@ -1,1 +1,0 @@
-venv/
