--- transforms/code/comment_cleanser/python/test-data/expected/metadata.json
@@ -1,1 +1,0 @@
-{"Removed code comment line": 189}
