--- transforms/code/comment_cleanser/ray/.dockerignore
@@ -1,1 +1,0 @@
-venv/
