--- transforms/code/comment_cleanser/ray/test-data/expected/metadata.json
@@ -1,57 +1,0 @@
-{
-  "pipeline": "pipeline_id",
-  "job details": {
-    "job category": "preprocessing",
-    "job name": "comment_cleanser",
-    "job type": "ray",
-    "job id": "job_id",
-    "start_time": "2025-03-17 19:56:03",
-    "end_time": "2025-03-17 19:56:09",
-    "status": "success"
-  },
-  "code": null,
-  "job_input_params": {
-    "contents_column_name": "contents",
-    "n_processes": 5,
-    "timeout": 300,
-    "skip_timeout": false,
-    "document_id_column_name": "document_id",
-    "checkpointing": false,
-    "max_files": -1,
-    "random_samples": -1,
-    "files_to_use": [
-      ".parquet"
-    ],
-    "number of workers": 1,
-    "worker options": {
-      "num_cpus": 0.8,
-      "max_restarts": -1
-    },
-    "actor creation delay": 0
-  },
-  "execution_stats": {
-    "cpus": 16,
-    "gpus": 0,
-    "memory": 5.832000732421875,
-    "object_store": 2.9160003662109375,
-    "execution time, min": 0.094
-  },
-  "job_output_stats": {
-    "source_files": 1,
-    "source_size": 16434,
-    "result_files": 1,
-    "result_size": 13483,
-    "processing_time": 4.573,
-    "Removed code comment line": 189,
-    "source_doc_count": 11,
-    "result_doc_count": 11
-  },
-  "source": {
-    "name": "/home/yash/dead_code/data-prep-kit/transforms/code/comment_cleanser/ray/test-data/input",
-    "type": "path"
-  },
-  "target": {
-    "name": "/home/yash/dead_code/data-prep-kit/transforms/code/comment_cleanser/ray/test-data/expected",
-    "type": "path"
-  }
-}
