+++ transforms/code/comment_cleanser/test-data/expected/metadata.json
@@ -1,0 +1,1 @@
+{"Removed code comment line": 189}
